refactor: правки код ревью
