--- src/components/library-filter/library-filter.stories.tsx
@@ -1,16 +1,0 @@
-import { ComponentStory, ComponentMeta } from '@storybook/react';
-import LibraryFilter from './library-filter';
-
-export default {
-  title: 'Components/LibraryFilter',
-  component: LibraryFilter
-} as ComponentMeta<typeof LibraryFilter>;
-
-const Template: ComponentStory<typeof LibraryFilter> = (args) => <LibraryFilter {...args}/>;
-
-export const Default = Template.bind({});
-Default.args = {
-  years: Array.from({ length: 8 }, () => ({ value: 2022, text: '2022' })),
-  programmes: Array.from({ length: 4 }, () => ({ pk: 2, name: 'Шорт-лист' }))
-};
-
