+++ src/components/ui/blog-card/blog-card.tsx
@@ -42,7 +42,6 @@ export const BlogCard: FC<BlogCardProps> = (props) => {
               objectFit="cover"
               priority
               onLoadingComplete={handleImageLoad}
-              unoptimized={Boolean(process.env.STORYBOOK)}
             />
           </div>
           <div className={styles.content}>
